requirements3.txt
adal==1.2.7
azure-applicationinsights==0.1.0
azure-batch==4.1.3
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
azure-datalake-store==0.0.52
azure-eventgrid==1.3.0
azure-graphrbac==0.40.0
azure-keyvault==1.1.0
azure-loganalytics==0.1.0
azure-mgmt==4.0.0
azure-mgmt-advisor==1.0.1
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.50.0
azure-mgmt-batch==5.0.1
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-cdn==3.1.0
azure-mgmt-cognitiveservices==3.0.0
azure-mgmt-commerce==1.0.1
azure-mgmt-compute==4.6.2
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-containerservice==4.4.0
azure-mgmt-cosmosdb==0.4.1
azure-mgmt-datafactory==0.6.0
azure-mgmt-datalake-analytics==0.6.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==1.0.0
azure-mgmt-devspaces==0.1.0
azure-mgmt-devtestlabs==2.2.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==1.0.0
azure-mgmt-eventhub==2.6.0
azure-mgmt-hanaonazure==0.1.1
azure-mgmt-iotcentral==0.1.0
azure-mgmt-iothub==0.5.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-loganalytics==0.2.0
azure-mgmt-logic==3.0.0
azure-mgmt-machinelearningcompute==0.4.1
azure-mgmt-managementgroups==0.1.0
azure-mgmt-managementpartner==0.1.1
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.1.0
azure-mgmt-media==1.0.0
azure-mgmt-monitor==0.5.2
azure-mgmt-msi==0.2.0
azure-mgmt-network==2.7.0
azure-mgmt-notificationhubs==2.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.1.0
azure-mgmt-powerbiembedded==2.0.0
azure-mgmt-rdbms==1.9.0
azure-mgmt-recoveryservices==0.3.0
azure-mgmt-recoveryservicesbackup==0.3.0
azure-mgmt-redis==5.0.0
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.2.1
azure-mgmt-resource==2.2.0
azure-mgmt-scheduler==2.0.0
azure-mgmt-search==2.1.0
azure-mgmt-servicebus==0.5.3
azure-mgmt-servicefabric==0.2.0
azure-mgmt-signalr==0.1.1
azure-mgmt-sql==0.9.1
azure-mgmt-storage==2.0.0
azure-mgmt-subscription==0.2.0
azure-mgmt-trafficmanager==0.50.0
azure-mgmt-web==0.35.0
azure-nspkg==3.0.2
azure-servicebus==0.21.1
azure-servicefabric==6.3.0.0
azure-servicemanagement-legacy==0.20.7
azure-storage-blob==1.5.0
azure-storage-common==1.4.2
azure-storage-file==1.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
debugpy==1.4.3
decorator==5.0.9
deprecation==2.1.0
distlib==0.3.2
entrypoints==0.3
filelock==3.0.12
idna==2.10
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
isodate==0.6.0
janus==0.4.0
jedi==0.18.0
jupyter-client==7.0.2
jupyter-core==4.7.1
lxml==4.6.3
matplotlib-inline==0.1.3
msrest==0.6.21
msrestazure==0.6.4
nest-asyncio==1.5.1
oauthlib==3.1.1
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
platformdirs==2.3.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.2.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
six==1.16.0
tornado==6.1
traitlets==5.1.0
urllib3==1.26.6
virtualenv==20.8.0